split param castable to string tests for each rule
