+++ tests/PHPStan/Rules/Functions/data/implode-param-castable-to-string-functions-enum.php
@@ -1,0 +1,13 @@
+<?php declare(strict_types = 1); // lint >= 8.1
+
+namespace ImplodeParamCastableToStringFunctionsEnum;
+
+enum FooEnum
+{
+	case A;
+}
+
+function invalidUsages()
+{
+	implode(',', [FooEnum::A]);
+}
